Обновлены энпоинты и модели ответов при обновлении финансовых отчетов.
