+++ app/dependencies/__init__.py
@@ -7,7 +7,6 @@ from .net_profit import get_net_profit_service, get_net_profit_repository
 from .percent_by_tax import get_percent_by_tax_service, get_percent_by_tax_repository
 from .stocks_quantity import get_stocks_quantity_service, get_stocks_quantity_repository, validate_edit_quantity_data
 from .turnover import get_turnover_service, get_turnover_repository
-from .fin_reports import get_fin_reports_service
 from .penalties import get_penalty_service
 from .sales import get_sale_service
 from .filters import get_dates_period_filter, get_months_filter
@@ -50,7 +49,6 @@ __all__ = [
     "get_stocks_quantity_repository",
     "get_turnover_service",
     "get_turnover_repository",
-    "get_fin_reports_service",
     "get_dates_period_filter",
     "get_months_filter",
     "get_penalty_service",
