--- app/infrastructure/WildberriesAPI/fin_reports/__init__.py
@@ -1,9 +1,0 @@
-from .fin_reports import WBFinReportFetcher
-from .normalizer import FIELD_TYPES, normalize_wb_value
-
-
-__all__ = [
-    "WBFinReportFetcher",
-    "FIELD_TYPES",
-    "normalize_wb_value",
-]
